+++ src/Oro/Bundle/CMSBundle/Migrations/Schema/OroCMSBundleInstaller.php
@@ -30,13 +30,11 @@ class OroCMSBundleInstaller implements
 {
     use AttachmentExtensionAwareTrait;
 
-    const CMS_LOGIN_PAGE_TABLE = 'oro_cms_login_page';
-    const MAX_LOGO_IMAGE_SIZE_IN_MB = 10;
-    const MAX_BACKGROUND_IMAGE_SIZE_IN_MB = 10;
-    const MAX_IMAGE_SLIDE_MAIN_IMAGE_SIZE_IN_MB = 10;
-    const MAX_IMAGE_SLIDE_MEDIUM_IMAGE_SIZE_IN_MB = 10;
-    const MAX_IMAGE_SLIDE_SMALL_IMAGE_SIZE_IN_MB = 10;
-    const MAX_CONTENT_TEMPLATE_PREVIEW_IMAGE_SIZE_IN_MB = 10;
+    public const CMS_LOGIN_PAGE_TABLE = 'oro_cms_login_page';
+    public const MAX_LOGO_IMAGE_SIZE_IN_MB = 10;
+    public const MAX_BACKGROUND_IMAGE_SIZE_IN_MB = 10;
+    public const MAX_CONTENT_TEMPLATE_PREVIEW_IMAGE_SIZE_IN_MB = 10;
+    public const MAX_SLIDE_IMAGE_SIZE_IN_MB = 10;
 
     protected ExtendExtension $extendExtension;
     protected SlugExtension $slugExtension;
@@ -46,7 +44,7 @@ class OroCMSBundleInstaller implements
      */
     public function getMigrationVersion(): string
     {
-        return 'v1_14';
+        return 'v1_14_1';
     }
 
     /**
@@ -468,31 +466,23 @@ class OroCMSBundleInstaller implements
         $table->addColumn('slide_order', 'integer', ['default' => 0]);
         $table->addColumn('url', 'string', ['length' => 255]);
         $table->addColumn('display_in_same_window', 'boolean', ['default' => true]);
-        $table->addColumn('title', 'string', ['length' => 255]);
+        $table->addColumn('alt_image_text', 'string', ['length' => 255]);
+        $table->addColumn('header', 'string', ['length' => 255, 'notnull' => false]);
         $table->addColumn('text', 'text', ['notnull' => false]);
         $table->addColumn('text_alignment', 'string', ['length' => 20, 'default' => ImageSlide::TEXT_ALIGNMENT_CENTER]);
 
-        $this->attachmentExtension->addImageRelation(
-            $schema,
-            'oro_cms_image_slide',
-            'mainImage',
-            ['attachment' => ['acl_protected' => false, 'use_dam' => true]],
-            self::MAX_IMAGE_SLIDE_MAIN_IMAGE_SIZE_IN_MB
-        );
-        $this->attachmentExtension->addImageRelation(
-            $schema,
-            'oro_cms_image_slide',
-            'mediumImage',
-            ['attachment' => ['acl_protected' => false, 'use_dam' => true]],
-            self::MAX_IMAGE_SLIDE_MEDIUM_IMAGE_SIZE_IN_MB
-        );
-        $this->attachmentExtension->addImageRelation(
-            $schema,
-            'oro_cms_image_slide',
-            'smallImage',
-            ['attachment' => ['acl_protected' => false, 'use_dam' => true]],
-            self::MAX_IMAGE_SLIDE_SMALL_IMAGE_SIZE_IN_MB
-        );
+        $this->addSlideImageRelation($schema, 'extraLargeImage');
+        $this->addSlideImageRelation($schema, 'extraLargeImage2x');
+        $this->addSlideImageRelation($schema, 'extraLargeImage3x');
+        $this->addSlideImageRelation($schema, 'largeImage');
+        $this->addSlideImageRelation($schema, 'largeImage2x');
+        $this->addSlideImageRelation($schema, 'largeImage3x');
+        $this->addSlideImageRelation($schema, 'mediumImage');
+        $this->addSlideImageRelation($schema, 'mediumImage2x');
+        $this->addSlideImageRelation($schema, 'mediumImage3x');
+        $this->addSlideImageRelation($schema, 'smallImage');
+        $this->addSlideImageRelation($schema, 'smallImage2x');
+        $this->addSlideImageRelation($schema, 'smallImage3x');
 
         $table->addColumn('organization_id', 'integer', ['notnull' => false]);
         $table->setPrimaryKey(['id']);
@@ -761,6 +751,17 @@ class OroCMSBundleInstaller implements
             ['organization_id'],
             ['id'],
             ['onDelete' => 'SET NULL']
+        );
+    }
+
+    public function addSlideImageRelation(Schema $schema, string $sourceColumnName): void
+    {
+        $this->attachmentExtension->addImageRelation(
+            $schema,
+            'oro_cms_image_slide',
+            $sourceColumnName,
+            ['attachment' => ['acl_protected' => false, 'use_dam' => true]],
+            self::MAX_SLIDE_IMAGE_SIZE_IN_MB
         );
     }