+++ CHANGELOG.md
@@ -77,6 +77,10 @@ The current file describes significant changes in the code that may affect the u
 
 #### CMSBundle
 * Added a new localizable `labels` field to `\Oro\Bundle\CMSBundle\Entity\ContentWidget`.
+* Added new image fields to `\Oro\Bundle\CMSBundle\Entity\ImageSlide`: `extraLargeImage2x`,`extraLargeImage3x`,`largeImage`,`largeImage2x`,`largeImage3x`,`mediumImage2x`,`mediumImage3x`,`smallImage2x`,`smallImage3x` .
+* Added a new `header` field to `\Oro\Bundle\CMSBundle\Entity\ImageSlide`.
+* Added a new twig function `oro_cms_image_slide_image` that returns the URL to the slider image using a predefined fallback strategy.
+* Added the `Oro\Bundle\CMSBundle\Validator\Constraints\HasAtLeastOneSizeImage` constraint and `\Oro\Bundle\CMSBundle\Validator\Constraints\HasAtLeastOneSizeImageValidator` validator to ensure at least one size of image is selected in a group.
 
 ### Changed
 
@@ -121,6 +125,12 @@ The current file describes significant changes in the code that may affect the u
 * Decomposed `\Oro\Bundle\ProductBundle\Model\ProductKitItemLineItemInterface` into `\Oro\Bundle\ProductBundle\Model\ProductKitItemAwareInterface`.
 * Changed `\Oro\Bundle\ProductBundle\ImportExport\Strategy\ProductStrategy` - added the possibility of postponing rows for missing product kits item products and fixed losing the related kitItems collections located deeper than 1st level.
 * Changed `\Oro\Bundle\ProductBundle\ImportExport\TemplateFixture\ProductFixture` - added new simple and kit products.
+* Field names were changed for form type `\Oro\Bundle\ProductBundle\Form\Type\ProductSegmentContentWidgetSettingsType`
+  - field `enable_autoplay` renamed to `autoplay`
+  - field `autoplay_speed` renamed to `autoplaySpeed`
+  - field `show_dots` renamed to `dots`
+  - field `enable_infinite_scroll` renamed to `infinite`
+  - field `autoplay_speed` renamed to `autoplaySpeed`
 
 #### SaleBundle
 * Changed `\Oro\Bundle\SaleBundle\Quote\Shipping\LineItem\Converter\FirstOffers\FirstOffersQuoteToShippingLineItemConverter` and `\Oro\Bundle\SaleBundle\Quote\Shipping\LineItem\Converter\SelectedOffers\SelectedOffersQuoteToShippingLineItemConverter` to make them work with the new shipping line item factory `\Oro\Bundle\SaleBundle\Quote\Shipping\Context\LineItem\Factory\ShippingLineItemFromQuoteProductDemandFactory`. 
@@ -135,6 +145,10 @@ The current file describes significant changes in the code that may affect the u
 #### ShoppingListBundle
 * Made use of `\Symfony\Component\Validator\Constraints\GroupSequence` in `\Oro\Bundle\ShoppingListBundle\Controller\Frontend\AjaxProductKitLineItemController` when creating/updating a product kit line item.
 * Made use of nested validation groups transforming into `\Symfony\Component\Validator\Constraints\GroupSequence` for `shoppinglistitem` and `shoppinglistkititem` storefront API resources.
+
+#### CMSBundle
+* Renamed field `title` to `altImageText`.
+* Renamed field `mainImage` to `extraLargeImage`.
 
 ### Removed